requirements.txt
streamlit==1.36.0
streamlit
PyPDF2
langchain-google-genai
google-generativeai
langchain-community
langchain
chromadb